feat: Add utility function for merging CSS class names
This commit adds a utility function `cn` to the `utils.ts` file. The `cn` function takes multiple class names as arguments and merges them using the `clsx` and `twMerge` libraries. This function will be useful for combining CSS class names in a more concise and efficient way.

Note: This commit message follows the conventional format of starting with a type (feat) followed by a colon and a brief description of the changes.